Ajustando de int para float
Corrigido.

+++ challenges/3-area-trapezio/c++/feliper2002/trapezio.cpp
@@ -3,13 +3,13 @@
 
 using namespace std;
 
-void areaTrapezio(int baseMaior, int baseMenor, int altura) {
+void areaTrapezio(float baseMaior, float baseMenor, float altura) {
     float area = ((baseMaior+baseMenor)*altura)/2;
     printf("A area do trapezio e: %.1f", area);
 }
 
 int main() {
-    int baseMaior, baseMenor, altura;
+    float baseMaior, baseMenor, altura;
     cout << "Base menor: "; cin >> baseMenor;
     cout << "Base maior: "; cin >> baseMaior;
     cout << "Altura: "; cin >> altura;
